captcha: Extract the Altcha helpers into a new zerver/lib/captcha.py.
